+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.9.2 Unreleased
 
+- Rename input/output json fields.
 - Indicate in CLI usage that container name is optional.
 
 # 0.9.1 2020.06.08